pyproject.toml
[tool.poetry]
name = "my-tg-bot"
version = "0.1.0"
description = ""
authors = ["AntonLysachev <anton@lysachev.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
Flask = "^3.0.2"
pyTelegramBotAPI = "^4.16.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
